+++ Events.php
@@ -9,8 +9,10 @@
 
 namespace humhub\modules\altNotification;
 
+use humhub\modules\notification\widgets\NotificationSettingsForm;
 use humhub\modules\space\MemberEvent;
 use Yii;
+use yii\base\Event;
 
 class Events
 {
@@ -23,13 +25,28 @@ class Events
             return;
         }
 
-        /** @var Module $module */
-        $module = Yii::$app->getModule('alt-notification');
-        $spaceGuids = $module->configuration->newContentNotifSpaceGuids;
+        $module = Module::getInstance();
 
         // If the Space is in **Module Settings**, auto-add it to their **User Settings**.
-        if (in_array($space->guid, $spaceGuids, true)) {
+        if (in_array($space->guid, $module->configuration->getNewContentNotifSpaceGuids(), true)) {
             Yii::$app->notification->setSpaceSetting($user, $space, true);
         }
+    }
+
+    public static function onNotificationSettingsFormBeforeRun(Event $event): void
+    {
+        /** @var NotificationSettingsForm $form */
+        $form = $event->sender;
+
+        if ($form->model->user) {
+            $module = Module::getInstance();
+            if (!$module->configuration->notifyForAllSpaces) {
+                // Only hide for admin settings, not for user settings
+                return;
+            }
+        }
+
+        // Hide the Spaces picker default notification on new content created in the notification settings form, as it is replaced by the module configuration.
+        $form->showSpaces = false;
     }
 }