+++ messages/ar/config.php
@@ -1,6 +1,6 @@
 <?php
 return [
   'Select Spaces for which Users should be notified about new content upon becoming a member.' => 'حدد المساحات التي يجب إخطار المستخدمين بالمحتوى الجديد فيها عند أن يصبحوا أعضاءً.',
-  'This module replaces the "{fieldName}" Notification settings, which means that you can leave it empty.' => 'تحل هذه الوحدة محل إعدادات الإشعار " {fieldName} "، مما يعني أنه يمكنك تركها فارغة.',
+  'This module replaces the "{fieldName}" Notification settings, which is hidden by this module.' => 'تستبدل هذه الوحدة إعدادات الإشعارات " {fieldName} "، والتي يتم إخفاؤها بواسطة هذه الوحدة.',
   'When a user joins a Space, if it is in this list, it is added to their "{fieldName}" Notification settings.' => 'عندما ينضم مستخدم إلى مساحة، إذا كان موجودًا في هذه القائمة، فسيتم إضافته إلى إعدادات الإشعار " {fieldName} " الخاصة به.',
 ];