+++ messages/br/config.php
@@ -2,6 +2,6 @@
 
 return [
     'Select Spaces for which Users should be notified about new content upon becoming a member.' => '',
-    'This module replaces the "{fieldName}" Notification settings, which means that you can leave it empty.' => '',
+    'This module replaces the "{fieldName}" Notification settings, which is hidden by this module.' => '',
     'When a user joins a Space, if it is in this list, it is added to their "{fieldName}" Notification settings.' => '',
 ];