+++ messages/cs/config.php
@@ -1,6 +1,6 @@
 <?php
 return [
   'Select Spaces for which Users should be notified about new content upon becoming a member.' => 'Vyberte prostory, u kterých by uživatelé měli být upozorněni na nový obsah, jakmile se stanou členem.',
-  'This module replaces the "{fieldName}" Notification settings, which means that you can leave it empty.' => 'Tento modul nahrazuje nastavení upozornění " {fieldName} ", což znamená, že jej můžete nechat prázdné.',
+  'This module replaces the "{fieldName}" Notification settings, which is hidden by this module.' => 'Tento modul nahrazuje nastavení oznámení „ {fieldName} “, které je tímto modulem skryté.',
   'When a user joins a Space, if it is in this list, it is added to their "{fieldName}" Notification settings.' => 'Když se uživatel připojí k prostoru, pokud je v tomto seznamu, přidá se do jeho nastavení oznámení " {fieldName} ".',
 ];