+++ messages/cy/config.php
@@ -1,6 +1,6 @@
 <?php
 return [
   'Select Spaces for which Users should be notified about new content upon becoming a member.' => 'Dewiswch Mannau y dylid hysbysu Defnyddwyr ar eu cyfer am gynnwys newydd wrth ddod yn aelod.',
-  'This module replaces the "{fieldName}" Notification settings, which means that you can leave it empty.' => 'Mae\'r modiwl hwn yn disodli\'r gosodiadau Hysbysiad " {fieldName} " , sy\'n golygu y gallwch ei adael yn wag.',
+  'This module replaces the "{fieldName}" Notification settings, which is hidden by this module.' => 'Mae\'r modiwl hwn yn disodli\'r gosodiadau Hysbysu " {fieldName} ", sydd wedi\'u cuddio gan y modiwl hwn.',
   'When a user joins a Space, if it is in this list, it is added to their "{fieldName}" Notification settings.' => 'Pan fydd defnyddiwr yn ymuno â Gofod, os yw yn y rhestr hon, mae\'n cael ei ychwanegu at ei osodiadau Hysbysiad {fieldName} .',
 ];