+++ messages/da/config.php
@@ -1,6 +1,6 @@
 <?php
 return [
   'Select Spaces for which Users should be notified about new content upon becoming a member.' => 'Vælg rum, som brugere skal have besked om nyt indhold for, når de bliver medlem.',
-  'This module replaces the "{fieldName}" Notification settings, which means that you can leave it empty.' => 'Dette modul erstatter " {fieldName} "-meddelelsesindstillingerne, hvilket betyder, at du kan lade det stå tomt.',
+  'This module replaces the "{fieldName}" Notification settings, which is hidden by this module.' => 'Dette modul erstatter notifikationsindstillingerne " {fieldName} ", som er skjult af dette modul.',
   'When a user joins a Space, if it is in this list, it is added to their "{fieldName}" Notification settings.' => 'Når en bruger tilmelder sig et space, hvis det er på denne liste, føjes det til deres " {fieldName} "-meddelelsesindstillinger.',
 ];