+++ messages/el/config.php
@@ -1,6 +1,6 @@
 <?php
 return [
   'Select Spaces for which Users should be notified about new content upon becoming a member.' => 'Επιλέξτε Χώρους για τους οποίους οι χρήστες θα πρέπει να ειδοποιούνται για νέο περιεχόμενο μόλις γίνουν μέλη.',
-  'This module replaces the "{fieldName}" Notification settings, which means that you can leave it empty.' => 'Αυτή η ενότητα αντικαθιστά τις ρυθμίσεις ειδοποίησης " {fieldName} ", πράγμα που σημαίνει ότι μπορείτε να το αφήσετε κενό.',
+  'This module replaces the "{fieldName}" Notification settings, which is hidden by this module.' => 'Αυτή η ενότητα αντικαθιστά τις ρυθμίσεις ειδοποιήσεων " {fieldName} ", οι οποίες είναι κρυφές από αυτήν την ενότητα.',
   'When a user joins a Space, if it is in this list, it is added to their "{fieldName}" Notification settings.' => 'Όταν ένας χρήστης συμμετέχει σε ένα Space, εάν βρίσκεται σε αυτήν τη λίστα, προστίθεται στις ρυθμίσεις ειδοποίησης του " {fieldName} ".',
 ];