+++ messages/fi/config.php
@@ -1,6 +1,6 @@
 <?php
 return [
   'Select Spaces for which Users should be notified about new content upon becoming a member.' => 'Valitse tilat, joissa käyttäjille ilmoitetaan uudesta sisällöstä heidän liittyessään.',
-  'This module replaces the "{fieldName}" Notification settings, which means that you can leave it empty.' => 'Tämä moduuli korvaa " {fieldName} " Ilmoitusasetukset, mikä tarkoittaa, että voit jättää sen tyhjäksi.',
+  'This module replaces the "{fieldName}" Notification settings, which is hidden by this module.' => 'Tämä moduuli korvaa ilmoitusasetukset " {fieldName} ", jotka tämä moduuli piilottaa.',
   'When a user joins a Space, if it is in this list, it is added to their "{fieldName}" Notification settings.' => 'Kun käyttäjä liittyy Space-ryhmään, jos se on tässä luettelossa, se lisätään hänen " {fieldName} " -ilmoitusasetuksiinsa.',
 ];