+++ messages/he/config.php
@@ -1,6 +1,6 @@
 <?php
 return [
   'Select Spaces for which Users should be notified about new content upon becoming a member.' => 'בחר מרחבים שעבורם משתמשים צריכים לקבל הודעה על תוכן חדש עם הצטרפותם לחברים.',
-  'This module replaces the "{fieldName}" Notification settings, which means that you can leave it empty.' => 'מודול זה מחליף את הגדרות ההתראה " {fieldName} ", מה שאומר שאתה יכול להשאיר אותו ריק.',
+  'This module replaces the "{fieldName}" Notification settings, which is hidden by this module.' => 'מודול זה מחליף את הגדרות ההתראות " {fieldName} ", אשר מוסתרות על ידי מודול זה.',
   'When a user joins a Space, if it is in this list, it is added to their "{fieldName}" Notification settings.' => 'כאשר משתמש מצטרף למרחב, אם הוא ברשימה זו, הוא מתווסף להגדרות ההתראה שלו " {fieldName} ".',
 ];