+++ messages/ht/config.php
@@ -1,6 +1,6 @@
 <?php
 return [
   'Select Spaces for which Users should be notified about new content upon becoming a member.' => 'Chwazi Espas pou Itilizatè yo ta dwe avèti sou nouvo kontni lè yo vin yon manm.',
-  'This module replaces the "{fieldName}" Notification settings, which means that you can leave it empty.' => 'Modil sa a ranplase paramèt Notifikasyon " {fieldName} ", ki vle di ou ka kite li vid.',
+  'This module replaces the "{fieldName}" Notification settings, which is hidden by this module.' => 'Modil sa a ranplase paramèt notifikasyon " {fieldName} " yo, ki kache pa modil sa a.',
   'When a user joins a Space, if it is in this list, it is added to their "{fieldName}" Notification settings.' => 'Lè yon itilizatè rantre nan yon Espas, si li nan lis sa a, li ajoute nan paramèt Notifikasyon " {fieldName} " yo.',
 ];