+++ messages/it/config.php
@@ -1,6 +1,6 @@
 <?php
 return [
   'Select Spaces for which Users should be notified about new content upon becoming a member.' => 'Seleziona gli spazi per i quali gli utenti devono essere informati dei nuovi contenuti una volta diventati membri.',
-  'This module replaces the "{fieldName}" Notification settings, which means that you can leave it empty.' => 'Questo modulo sostituisce le impostazioni di notifica " {fieldName} ", il che significa che puoi lasciarlo vuoto.',
+  'This module replaces the "{fieldName}" Notification settings, which is hidden by this module.' => 'Questo modulo sostituisce le impostazioni di notifica " {fieldName} ", che vengono nascoste da questo modulo.',
   'When a user joins a Space, if it is in this list, it is added to their "{fieldName}" Notification settings.' => 'Quando un utente si unisce a uno spazio, se è presente in questo elenco, viene aggiunto alle sue impostazioni di notifica " {fieldName} ".',
 ];