+++ messages/ja/config.php
@@ -1,6 +1,6 @@
 <?php
 return [
   'Select Spaces for which Users should be notified about new content upon becoming a member.' => 'ユーザーがメンバーになったときに新しいコンテンツについて通知を受けるスペースを選択します。',
-  'This module replaces the "{fieldName}" Notification settings, which means that you can leave it empty.' => 'このモジュールは、「 {fieldName} 」通知設定を置き換えます。つまり、空のままにしておくことができます。',
+  'This module replaces the "{fieldName}" Notification settings, which is hidden by this module.' => 'このモジュールは、「 {fieldName} 」通知設定を置き換えます。このモジュールによって、通知設定は非表示になります。',
   'When a user joins a Space, if it is in this list, it is added to their "{fieldName}" Notification settings.' => 'ユーザーがスペースに参加すると、そのスペースがこのリストに含まれている場合、そのユーザーの「 {fieldName} 」通知設定に追加されます。',
 ];