+++ messages/lv/config.php
@@ -1,6 +1,6 @@
 <?php
 return [
   'Select Spaces for which Users should be notified about new content upon becoming a member.' => 'Atlasiet Vietas, par kurām lietotājiem jāinformē par jaunu saturu, kļūstot par dalībnieku.',
-  'This module replaces the "{fieldName}" Notification settings, which means that you can leave it empty.' => 'Šis modulis aizstāj paziņojumu iestatījumus " {fieldName} ", kas nozīmē, ka varat to atstāt tukšu.',
+  'This module replaces the "{fieldName}" Notification settings, which is hidden by this module.' => 'Šis modulis aizstāj paziņojumu iestatījumus " {fieldName} ", kas ir paslēpti šajā modulī.',
   'When a user joins a Space, if it is in this list, it is added to their "{fieldName}" Notification settings.' => 'Kad lietotājs pievienojas vietai, ja tā ir šajā sarakstā, tā tiek pievienota viņa " {fieldName} " paziņojumu iestatījumiem.',
 ];