+++ messages/nl/config.php
@@ -1,6 +1,6 @@
 <?php
 return [
   'Select Spaces for which Users should be notified about new content upon becoming a member.' => 'Selecteer de ruimtes waarvoor gebruikers op de hoogte moeten worden gesteld van nieuwe content zodra ze lid worden.',
-  'This module replaces the "{fieldName}" Notification settings, which means that you can leave it empty.' => 'Deze module vervangt de meldingsinstellingen " {fieldName} ", wat betekent dat u deze leeg kunt laten.',
+  'This module replaces the "{fieldName}" Notification settings, which is hidden by this module.' => 'Deze module vervangt de meldingsinstellingen van " {fieldName} ", die door deze module verborgen worden.',
   'When a user joins a Space, if it is in this list, it is added to their "{fieldName}" Notification settings.' => 'Wanneer een gebruiker zich bij een ruimte aansluit en deze in deze lijst staat, wordt deze toegevoegd aan de meldingsinstellingen van \' {fieldName} \'.',
 ];