+++ messages/pt-BR/config.php
@@ -1,6 +1,6 @@
 <?php
 return [
   'Select Spaces for which Users should be notified about new content upon becoming a member.' => 'Selecione Espaços para os quais os Usuários devem ser notificados sobre novos conteúdos ao se tornarem membros.',
-  'This module replaces the "{fieldName}" Notification settings, which means that you can leave it empty.' => 'Este módulo substitui as configurações de notificação " {fieldName} ", o que significa que você pode deixá-lo vazio.',
+  'This module replaces the "{fieldName}" Notification settings, which is hidden by this module.' => 'Este módulo substitui as configurações de notificação " {fieldName} ", que ficam ocultas por este módulo.',
   'When a user joins a Space, if it is in this list, it is added to their "{fieldName}" Notification settings.' => 'Quando um usuário entra em um Espaço, se ele estiver nesta lista, ele será adicionado às configurações de Notificação " {fieldName} ".',
 ];