+++ messages/ro/config.php
@@ -1,6 +1,6 @@
 <?php
 return [
   'Select Spaces for which Users should be notified about new content upon becoming a member.' => 'Selectați Spații pentru care Utilizatorii ar trebui să fie anunțați despre conținutul nou atunci când devin membru.',
-  'This module replaces the "{fieldName}" Notification settings, which means that you can leave it empty.' => 'Acest modul înlocuiește setările de notificare „ {fieldName} ”, ceea ce înseamnă că îl puteți lăsa necompletat.',
+  'This module replaces the "{fieldName}" Notification settings, which is hidden by this module.' => 'Acest modul înlocuiește setările de notificare „ {fieldName} ”, care sunt ascunse de acest modul.',
   'When a user joins a Space, if it is in this list, it is added to their "{fieldName}" Notification settings.' => 'Când un utilizator se alătură unui spațiu, dacă acesta se află în această listă, acesta este adăugat la setările sale de notificare „ {fieldName} ”.',
 ];