+++ messages/sr/config.php
@@ -1,6 +1,6 @@
 <?php
 return [
   'Select Spaces for which Users should be notified about new content upon becoming a member.' => 'Изаберите просторе за које би корисници требало да буду обавештени о новом садржају када постану члан.',
-  'This module replaces the "{fieldName}" Notification settings, which means that you can leave it empty.' => 'Овај модул замењује подешавања обавештења „ {fieldName} “, што значи да га можете оставити празним.',
+  'This module replaces the "{fieldName}" Notification settings, which is hidden by this module.' => 'Овај модул замењује подешавања обавештења „ {fieldName} “, која су скривена овим модулом.',
   'When a user joins a Space, if it is in this list, it is added to their "{fieldName}" Notification settings.' => 'Када се корисник придружи простору, ако се налази на овој листи, додаје се у подешавања обавештења „ {fieldName} “.',
 ];