+++ messages/sw/config.php
@@ -1,6 +1,6 @@
 <?php
 return [
   'Select Spaces for which Users should be notified about new content upon becoming a member.' => 'Chagua Nafasi ambazo Watumiaji wanapaswa kuarifiwa kuhusu maudhui mapya baada ya kuwa wanachama.',
-  'This module replaces the "{fieldName}" Notification settings, which means that you can leave it empty.' => 'Moduli hii inachukua nafasi ya " {fieldName} " Mipangilio ya Arifa, ambayo ina maana kwamba unaweza kuiacha tupu.',
+  'This module replaces the "{fieldName}" Notification settings, which is hidden by this module.' => 'Moduli hii inachukua nafasi ya mipangilio ya arifa ya " {fieldName} ", ambayo imefichwa na moduli hii.',
   'When a user joins a Space, if it is in this list, it is added to their "{fieldName}" Notification settings.' => 'Mtumiaji anapojiunga na Nafasi, ikiwa iko katika orodha hii, inaongezwa kwa mipangilio yake ya Arifa ya " {fieldName} ".',
 ];