+++ messages/uk/config.php
@@ -1,6 +1,6 @@
 <?php
 return [
   'Select Spaces for which Users should be notified about new content upon becoming a member.' => 'Виберіть простори, для яких користувачі повинні отримувати сповіщення про новий вміст, коли стають учасниками.',
-  'This module replaces the "{fieldName}" Notification settings, which means that you can leave it empty.' => 'Цей модуль замінює параметри сповіщень " {fieldName} ", тобто ви можете залишити його порожнім.',
+  'This module replaces the "{fieldName}" Notification settings, which is hidden by this module.' => 'Цей модуль замінює налаштування сповіщень " {fieldName} ", які приховані цим модулем.',
   'When a user joins a Space, if it is in this list, it is added to their "{fieldName}" Notification settings.' => 'Коли користувач приєднується до чат-кімнати, якщо він є в цьому списку, він додається до його налаштувань сповіщень " {fieldName} ".',
 ];