+++ messages/uz/config.php
@@ -1,6 +1,6 @@
 <?php
 return [
   'Select Spaces for which Users should be notified about new content upon becoming a member.' => 'Foydalanuvchilar a\'zo bo\'lgandan keyin yangi kontent haqida xabardor qilinishi kerak bo\'lgan joylarni tanlang.',
-  'This module replaces the "{fieldName}" Notification settings, which means that you can leave it empty.' => 'Ushbu modul " {fieldName} " bildirishnoma sozlamalarini almashtiradi, ya\'ni siz uni bo\'sh qoldirishingiz mumkin.',
+  'This module replaces the "{fieldName}" Notification settings, which is hidden by this module.' => 'Ushbu modul ushbu modul tomonidan yashirilgan " {fieldName} " bildirishnoma sozlamalarini almashtiradi.',
   'When a user joins a Space, if it is in this list, it is added to their "{fieldName}" Notification settings.' => 'Agar foydalanuvchi bo\'shliqqa qo\'shilsa, u ushbu ro\'yxatda bo\'lsa, u " {fieldName} " bildirishnoma sozlamalariga qo\'shiladi.',
 ];