+++ messages/vi/config.php
@@ -1,6 +1,6 @@
 <?php
 return [
   'Select Spaces for which Users should be notified about new content upon becoming a member.' => 'Chọn Không gian mà Người dùng sẽ được thông báo về nội dung mới khi trở thành thành viên.',
-  'This module replaces the "{fieldName}" Notification settings, which means that you can leave it empty.' => 'Mô-đun này thay thế cài đặt Thông báo " {fieldName} ", nghĩa là bạn có thể để trống.',
+  'This module replaces the "{fieldName}" Notification settings, which is hidden by this module.' => 'Mô-đun này thay thế cài đặt Thông báo " {fieldName} ", cài đặt này bị ẩn đi bởi mô-đun này.',
   'When a user joins a Space, if it is in this list, it is added to their "{fieldName}" Notification settings.' => 'Khi người dùng tham gia một Không gian, nếu không gian đó nằm trong danh sách này, thì không gian đó sẽ được thêm vào cài đặt Thông báo " {fieldName} " của họ.',
 ];