+++ models/Configuration.php
@@ -10,6 +10,10 @@
 namespace humhub\modules\altNotification\models;
 
 use humhub\components\SettingsManager;
+use humhub\modules\notification\components\NotificationManager;
+use humhub\modules\space\models\Membership;
+use humhub\modules\space\models\Space;
+use humhub\modules\user\models\User;
 use Yii;
 use yii\base\Model;
 
@@ -18,6 +22,7 @@ class Configuration extends Model
     public SettingsManager $settingsManager;
 
     public array $newContentNotifSpaceGuids = [];
+    public bool $notifyForAllSpaces = false;
 
 
     /**
@@ -27,6 +32,7 @@ class Configuration extends Model
     {
         return [
             [['newContentNotifSpaceGuids'], 'safe'],
+            [['notifyForAllSpaces'], 'boolean'],
         ];
     }
 
@@ -37,6 +43,7 @@ class Configuration extends Model
     {
         return [
             'newContentNotifSpaceGuids' => Yii::t('AltNotificationModule.config', 'Select Spaces for which Users should be notified about new content upon becoming a member.'),
+            'notifyForAllSpaces' => Yii::t('AltNotificationModule.config', 'Notify Users for all Spaces'),
         ];
     }
 
@@ -47,15 +54,17 @@ class Configuration extends Model
     {
         return [
             'newContentNotifSpaceGuids'
-                => Yii::t('AltNotificationModule.config', 'When a user joins a Space, if it is in this list, it is added to their "{fieldName}" Notification settings.', [
-                    'fieldName' => Yii::t('NotificationModule.base', 'Receive \'New Content\' Notifications for the following spaces'),
-                ]),
+            => Yii::t('AltNotificationModule.config', 'When a user joins a Space, if it is in this list, it is added to their "{fieldName}" Notification settings.', [
+                'fieldName' => Yii::t('NotificationModule.base', 'By default, receive \'New Content\' Notifications for the following Spaces'),
+            ]),
+            'notifyForAllSpaces' => Yii::t('AltNotificationModule.config', 'By default, all Spaces receive \'New Content\' Notifications.'),
         ];
     }
 
     public function loadBySettings(): void
     {
         $this->newContentNotifSpaceGuids = (array)$this->settingsManager->getSerialized('newContentNotifSpaceGuids', $this->newContentNotifSpaceGuids);
+        $this->notifyForAllSpaces = (bool)$this->settingsManager->get('notifyForAllSpaces', $this->notifyForAllSpaces);
     }
 
     public function save(): bool
@@ -65,7 +74,37 @@ class Configuration extends Model
         }
 
         $this->settingsManager->setSerialized('newContentNotifSpaceGuids', $this->newContentNotifSpaceGuids);
+        $this->settingsManager->set('notifyForAllSpaces', $this->notifyForAllSpaces);
+
+        $this->updateAllUsersNotificationSettings();
 
         return true;
+    }
+
+    public function getNewContentNotifSpaceGuids()
+    {
+        return $this->notifyForAllSpaces
+            ? Space::find()
+                ->select('guid')
+                ->where(['status' => Space::STATUS_ENABLED])
+                ->column()
+            : $this->newContentNotifSpaceGuids;
+    }
+
+    private function updateAllUsersNotificationSettings()
+    {
+        // Add all Spaces the User is a member of, from **Module Settings**, to their **User Settings**.
+        /** @var User $user */
+        foreach (User::find()->active()->all() as $user) {
+            if (NotificationManager::isTouchedSettings($user)) {
+                continue;
+            }
+            $userSpaceMembershipGuids = Membership::find()
+                ->joinWith('space')
+                ->where(['user_id' => $user->id, 'space_membership.status' => Membership::STATUS_MEMBER])
+                ->select('space.guid')
+                ->column();
+            Yii::$app->notification->setSpaces(array_intersect($this->getNewContentNotifSpaceGuids(), $userSpaceMembershipGuids), $user);
+        }
     }
 }